'feat/master-data-controllers-and-models'

+++ app/Http/Controllers/BrandController.php
@@ -16,10 +16,22 @@ class BrandController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $brands = Brand::all();
-        return response()->json(['brands' => $brands]);
+        $page = $request->page;
+        $per_page = $request->per_page ?? 10000;
+        $search = $request->search;
+
+        $brand = Brand::when($search, function ($query, $search) {
+            return $query->where('brand_name', 'LIKE', '%' . $search . '%');
+        })
+            ->paginate($per_page, ['*'], 'page', $page);
+
+        return response()->json([
+            'status' => 'success',
+            'message' => 'Menampilkan data brand',
+            'brands' => $brand,
+        ], 200);
     }
 
     /**
@@ -47,31 +59,23 @@ class BrandController extends Controller
             ], 422);
         }
 
-        DB::startTransaction();
-        try {
-            $brand = Brand::create([
-                'brand_name' => $request->brand_name,
-            ]);
+        
+        $brand = Brand::create([
+            'brand_name' => ucwords($request->brand_name),
+            'user_id' => auth()->user()->id,
+        ]);
 
-            if (!$brand) {
-                throw new HttpException(400, 'Brand gagal ditambahkan');
-            }
-
-            DB::commit();
-
-            return response()->json([
-                'status' => 'success',
-                'message' => 'Brand berhasil ditambahkan',
-            ], 201);
-
-        } catch (HttpException $e) {
-            DB::rollBack();
-
+        if (!$brand) {
             return response()->json([
                 'status' => 'error',
-                'message' => $e->getMessage(),
-            ], $e->getStatusCode());
+                'message' => 'Brand gagal ditambahkan'
+            ], 400);
         }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => 'Brand berhasil ditambahkan',
+        ], 201);
     }
 
     /**
@@ -108,7 +112,8 @@ class BrandController extends Controller
         }
 
         $updateBrand = $brand->update([
-            'brand_name' => $request->brand_name,
+            'brand_name' => ucwords($request->brand_name),
+            'user_id' => auth()->user()->id,
         ]);
 
         if (!$updateBrand){
@@ -129,8 +134,50 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
-        $brand->delete();
+        $user = auth()->user();
 
-        return response()->json(['message' => 'Brand berhasil dihapus']);
+        if($user->role->name != 'Admin'){
+            $this->deactivate($brand->id);
+        }
+        else{
+            if ($brand->products()->exists()) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Tidak dapat menghapus data brand yang memiliki produk terkait'
+                ], 422);
+            }
+            if(!$brand->delete()){
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Gagal menghapus data brand',
+                ], 400);
+            }
+            $brand->delete();
+
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Berhasil menghapus data brand',
+            ]);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => $user->role->name != 'Admin' ? 'Berhasil mengonaktifkan data brand' : 'Berhasil menghapus data brand',
+        ]);
+    }
+
+    public function deactivate($id)
+    {
+        $updateBrand = Brand::where('id', $id)->update([
+            'is_active' => 0,
+            'deactivated_at' => Carbon::now(),
+        ]);
+
+        if(!$updateBrand){
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Gagal menonaktifkan brand',
+            ]);
+        }
     }
 }
